requirements.txt
psycopg2
Flask
Flask-SQLAlchemy
Flask-Migrate